+++ CHANGELOG.md
@@ -14,6 +14,9 @@ Versions before 0.1.0 are untracked
 * require sort import by isort
 
 
+### Changed
+* `pd ls` - `--column` is replaced by `--row`
+
 ## [0.1.3] - 2022-02-19
 ### Changed
 * move pypi publish from check to release workflow